OpenGL interaction is working, Renderable stream is implemented. Mesh implementation begun.

+++ src/material/phong.cl
@@ -1,4 +1,4 @@
-IN(uint4,color);
+IN(float4,color);
